refactor(chat-item): rename room-item to chat-item
