+++ frontend/src/app/components/chat/item/image.css.ts
